requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
werkzeug==3.0.1
psutil==5.9.6
cryptography==41.0.7
